% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hs_district.R
\docType{data}
\name{hs_district}
\alias{hs_district}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\references{
\url{data_blah.com}
}
\author{
CDC
}
\keyword{data}
