% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hs_substance_use.R
\docType{data}
\name{hs_substance_use}
\alias{hs_substance_use}
\title{This is dataset contains variables related to substance use among high school
students}
\description{
This is dataset contains variables related to substance use among high school
students
}
\references{
\url{data_blah.com}
}
\author{
CDC
}
\keyword{data}
