% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeTF.R
\name{RecodeTF}
\alias{RecodeTF}
\title{RecodeTF}
\usage{
RecodeTF(x)
}
\arguments{
\item{x}{The name of the vector you want to recode}
}
\value{
The vector will replace the numbers to a logical atomic vector
}
\description{
RecodeTF
}
