% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaleDays.R
\name{ScaleDays}
\alias{ScaleDays}
\title{Converting a scale that measures consumption of food from 1 to 4
times a day.
In the options they combine 7 days and "per day", therefore the
options including 7 days are averaged. Meaning, 1 to 3 times during the past
7 days is considered as 0.4, given that is 3 divided by 7.}
\usage{
ScaleDays(x)
}
\arguments{
\item{x}{The name of the vector to recode}
}
\value{
The values on the scale will change from factors to integers
}
\description{
Converting a scale that measures consumption of food from 1 to 4
times a day.
In the options they combine 7 days and "per day", therefore the
options including 7 days are averaged. Meaning, 1 to 3 times during the past
7 days is considered as 0.4, given that is 3 divided by 7.
}
