% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaleNotSure.R
\name{ScaleNotSure}
\alias{ScaleNotSure}
\title{Converting a scale with three options, to a logical TRUE-FALSE variable.
This scale has the oprions yes, no and not sure. Therefore no and not sure
will be FALSE and yes will be assigned TRUE.}
\usage{
ScaleNotSure(x)
}
\arguments{
\item{x}{The name of the vector to recode}
}
\value{
The values on the scale will change from factors to integers
}
\description{
Converting a scale with three options, to a logical TRUE-FALSE variable.
This scale has the oprions yes, no and not sure. Therefore no and not sure
will be FALSE and yes will be assigned TRUE.
}
