% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaleToNumber12.R
\name{ScaleToNumber12}
\alias{ScaleToNumber12}
\title{Converting a scale from 0 to 12 into  numeric values choosing the worst
case scenario. This means, option "2 or 3 days" will be translated as
3 days.}
\usage{
ScaleToNumber12(x)
}
\arguments{
\item{x}{The name of the vector to recode}
}
\value{
The values on the scale will change from factors to integers
}
\description{
Converting a scale from 0 to 12 into  numeric values choosing the worst
case scenario. This means, option "2 or 3 days" will be translated as
3 days.
}
