% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaleToNumber30.R
\name{ScaleToNumber30}
\alias{ScaleToNumber30}
\title{Converting a scale from 0 to 30 into  numeric values choosing the worst
case scenario. This means, option "6 to 9 days" will be translated as
9 days.}
\usage{
ScaleToNumber30(x)
}
\arguments{
\item{x}{The name of the vector to recode}
}
\value{
The values on the scale will change from factors to integers
}
\description{
Converting a scale from 0 to 30 into  numeric values choosing the worst
case scenario. This means, option "6 to 9 days" will be translated as
9 days.
}
