% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaleToNumber40.R
\name{ScaleToNumber40}
\alias{ScaleToNumber40}
\title{Converting a scale from 0 to 40 into  numeric values choosing the worst
case scenario. This means, option "20 to 39 times" will be translated as
39 times.}
\usage{
ScaleToNumber40(x)
}
\arguments{
\item{x}{The name of the vector to recode}
}
\value{
The values on the scale will change from factors to integers
}
\description{
Converting a scale from 0 to 40 into  numeric values choosing the worst
case scenario. This means, option "20 to 39 times" will be translated as
39 times.
}
