% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaleToNumber6.R
\name{ScaleToNumber6}
\alias{ScaleToNumber6}
\title{Converting a scale from 0 to 6 into  numeric values choosing the worst
case scenario. This means, option "2 to 3 times" will be tranlated as
3 times.}
\usage{
ScaleToNumber6(x)
}
\arguments{
\item{x}{The name of the vector to recode}
}
\value{
The values on the scale will change from factors to integers
}
\description{
Converting a scale from 0 to 6 into  numeric values choosing the worst
case scenario. This means, option "2 to 3 times" will be tranlated as
3 times.
}
