% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScalingToBinary12.R
\name{ScalingToBinary12}
\alias{ScalingToBinary12}
\title{Converting a scale from 0 to 12 into a binary variable}
\usage{
ScalingToBinary12(x)
}
\arguments{
\item{x}{The name of the vector to recode}
}
\value{
The values on the scale will change into a binary value of TRUE and FALSE
}
\description{
Converting a scale from 0 to 12 into a binary variable
}
