% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScalingToBinary30.R
\name{ScalingToBinary30}
\alias{ScalingToBinary30}
\title{Converting a scale from 0 to 30 into a binary variable}
\usage{
ScalingToBinary30(x)
}
\arguments{
\item{x}{the vector you want to recode}
}
\value{
The scale will change to a logical vector according to the scale.
}
\description{
Converting a scale from 0 to 30 into a binary variable
}
