% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScalingToBinary40.R
\name{ScalingToBinary40}
\alias{ScalingToBinary40}
\title{ScalingToBinary40}
\usage{
ScalingToBinary40(x)
}
\arguments{
\item{x}{The name of the vector you wan to recode}
}
\value{
Te vector will change to a logical atomic vector according to the values on the scale
}
\description{
ScalingToBinary40
}
