% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hs_district.R
\docType{data}
\name{hs_district}
\alias{hs_district}
\title{hs_district}
\format{
A tibble with 504, 249 rows and 314 variables
\url{https://www.cdc.gov/healthyyouth/data/yrbs/pdf/2019/2019_YRBS_SADC_Documentation.pdf}
}
\source{
<cdc.gov/healthyyouth/data/yrbs/pdf/2019/2019_YRBS_SADC_Dcumentation.pdf>
}
\usage{
data(hs_district)
}
\description{
This is dataset contains variables related to high school students'
demographics
}
\author{
Centers for Disease Control and Prevention, Youth Risk Behavior
Survey Data
}
\keyword{datasets}
