% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScalingToBinary40.R
\name{ScalingToBinary40}
\alias{ScalingToBinary40}
\title{Converting a scale from 0 to 40 into a binary variable}
\usage{
ScalingToBinary40(x)
}
\arguments{
\item{x}{The name of the vector you wan to recode}
}
\value{
Te vector will change to a logical atomic vector according to the values on the scale
}
\description{
Converting a scale from 0 to 40 into a binary variable
}
